fastAPI_app/requirements.txt
fastapi
uvicorn[standard]  
elasticsearch
openai
python-dotenv
sentence-transformers 
transformers==4.36.2
langchain   